feat: add configuration option to disable web configuration

+++ handlers/routes.go
@@ -7,14 +7,17 @@ import (
 	"github.com/nwesterhausen/domain-monitor/service"
 )
 
-func SetupRoutes(app *echo.Echo) {
-	app.GET("/", HandlerShowBase)
+func SetupRoutes(app *echo.Echo, includeConfiguration bool) {
+	bh := &BaseHandler{IncludeConfiguration: includeConfiguration}
+	app.GET("/", bh.HandlerShowBase)
 
 	app.GET("/dashboard", HandlerRenderDashboard)
-	app.GET("/configuration", HandlerRenderConfiguration)
+	if includeConfiguration {
+		app.GET("/configuration", HandlerRenderConfiguration)
+	}
 }
 
-func SetupDomainRoutes(app *echo.Echo, domains configuration.DomainConfiguration) {
+func SetupDomainRoutes(app *echo.Echo, domains configuration.DomainConfiguration, configurationEnabled bool) {
 	domainHtmx := app.Group("/domain")
 	domainApi := app.Group("/api/domain")
 
@@ -22,19 +25,23 @@ func SetupDomainRoutes(app *echo.Echo, domains configuration.DomainConfiguration
 	dhapi := NewApiDomainHandler(ds)
 	dh := NewDomainHandler(ds)
 
-	domainApi.POST("/create", dhapi.HandleDomainCreate)
 	domainApi.GET("", dhapi.HandleDomainList)
 	domainApi.GET("/:fqdn", dhapi.HandleDomainShow)
-	domainApi.PUT("/:fqdn", dhapi.HandleDomainUpdate)
-	domainApi.DELETE("/:fqdn", dhapi.HandleDomainDelete)
+	if configurationEnabled {
+		domainApi.POST("/create", dhapi.HandleDomainCreate)
+		domainApi.PUT("/:fqdn", dhapi.HandleDomainUpdate)
+		domainApi.DELETE("/:fqdn", dhapi.HandleDomainDelete)
+	}
 
 	domainHtmx.GET("/:fqdn/card", dh.GetCard)
 	domainHtmx.GET("/cards", dh.GetCards)
 	domainHtmx.GET("/tbody", dh.GetListTbody)
 	domainHtmx.GET("/edit/:fqdn", dh.GetEditDomain)
-	domainHtmx.POST("/update", dh.PostUpdateDomain)
-	domainHtmx.POST("/new", dh.PostNewDomain)
-	domainHtmx.DELETE("/:fqdn", dh.DeleteDomain)
+	if configurationEnabled {
+		domainHtmx.POST("/update", dh.PostUpdateDomain)
+		domainHtmx.POST("/new", dh.PostNewDomain)
+		domainHtmx.DELETE("/:fqdn", dh.DeleteDomain)
+	}
 }
 
 func SetupConfigRoutes(app *echo.Echo, config configuration.Configuration) {
@@ -45,13 +52,17 @@ func SetupConfigRoutes(app *echo.Echo, config configuration.Configuration) {
 	ch := NewConfigurationHandler(cs)
 
 	configApi.GET("/:section/:key", ch.GetSectionKey)
-	configApi.POST("/:section/:key", ch.SetSectionKey)
+	if config.Config.App.ShowConfiguration {
+		configApi.POST("/:section/:key", ch.SetSectionKey)
+	}
 
-	configGroup.GET("/app", ch.RenderAppConfiguration)
-	configGroup.GET("/domain", ch.RenderDomainConfiguration)
-	configGroup.GET("/smtp", ch.RenderSmtpConfiguration)
-	configGroup.GET("/scheduler", ch.RenderSchedulerConfiguration)
-	configGroup.GET("/alerts", ch.RenderAlertsConfiguration)
+	if config.Config.App.ShowConfiguration {
+		configGroup.GET("/app", ch.RenderAppConfiguration)
+		configGroup.GET("/domain", ch.RenderDomainConfiguration)
+		configGroup.GET("/smtp", ch.RenderSmtpConfiguration)
+		configGroup.GET("/scheduler", ch.RenderSchedulerConfiguration)
+		configGroup.GET("/alerts", ch.RenderAlertsConfiguration)
+	}
 }
 
 func SetupMailerRoutes(app *echo.Echo, ms *service.MailerService, alertRecipient string) {
